flatten classes
