+++ hobj/learning_models/__init__.py
@@ -3,7 +3,9 @@ from abc import ABC, abstractmethod
 
 import numpy as np
 
-from hobj.learning_models.linear_learning_model.representation import RepresentationalModel
+from hobj.learning_models.linear_learning_model.representation import (
+    RepresentationalModel,
+)
 from hobj.learning_models.linear_learning_model.update_rules import UpdateRule
 from hobj.types import ImageId
